requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
asyncpg==0.30.0
bcrypt==4.3.0
certifi==2025.6.15
click==8.2.1
colorama==0.4.6
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
et_xmlfile==2.0.0
fastapi==0.115.13
fastapi-cli==0.0.7
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
numpy==2.3.1
openpyxl==3.1.5
pandas==2.3.1
passlib==1.7.4
psutil==7.0.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyasn1==0.6.1
pydantic==2.11.7
pydantic-settings==2.10.0
pydantic_core==2.33.2
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
rich==14.0.0
rich-toolkit==0.14.7
rsa==4.9.1
schedule==1.2.2
scikit-learn==1.7.0
scipy==1.16.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
threadpoolctl==3.6.0
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
uvicorn==0.34.3
watchfiles==1.1.0
websockets==15.0.1
